+++ packages/twenty-sdk/src/cli/__tests__/apps/root-app/__integration__/app-dev/expected-manifest.ts
@@ -11,6 +11,7 @@ export const EXPECTED_MANIFEST: Manifest = {
     defaultRoleUniversalIdentifier: 'e1e2e3e4-e5e6-4000-8000-000000000002',
     packageJsonChecksum: '93ae1e2eb3db18351d06f43550700dcc',
     yarnLockChecksum: 'd41d8cd98f00b204e9800998ecf8427e',
+    apiClientChecksum: null,
   },
   publicAssets: [],
   fields: [],
