+++ packages/twenty-sdk/src/cli/utilities/dev/orchestrator/steps/upload-files-orchestrator-step.ts
@@ -3,7 +3,13 @@ import {
   type OrchestratorStateBuiltFileInfo,
 } from '@/cli/utilities/dev/orchestrator/dev-mode-orchestrator-state';
 import { FileUploader } from '@/cli/utilities/file/file-uploader';
-import { type FileFolder } from 'twenty-shared/types';
+import crypto from 'crypto';
+import * as fs from 'fs-extra';
+import { join } from 'path';
+import { OUTPUT_DIR } from 'twenty-shared/application';
+import { FileFolder } from 'twenty-shared/types';
+
+const API_CLIENT_FILES = ['types.ts', 'schema.ts'];
 
 export type UploadFilesOrchestratorStepOutput = {
   fileUploader: FileUploader | null;
@@ -101,6 +107,43 @@ export class UploadFilesOrchestratorStep {
 
     step.status = 'done';
     this.notify();
+  }
+
+  async copyAndUploadApiClientFiles(appPath: string): Promise<void> {
+    const generatedDir = join(
+      appPath,
+      'node_modules',
+      'twenty-sdk',
+      'generated',
+    );
+    const outputDir = join(appPath, OUTPUT_DIR, 'api-client');
+
+    await fs.ensureDir(outputDir);
+
+    for (const fileName of API_CLIENT_FILES) {
+      const absoluteSourcePath = join(generatedDir, fileName);
+
+      if (!(await fs.pathExists(absoluteSourcePath))) {
+        continue;
+      }
+
+      await fs.copy(absoluteSourcePath, join(outputDir, fileName));
+
+      const content = await fs.readFile(absoluteSourcePath);
+      const checksum = crypto.createHash('md5').update(content).digest('hex');
+
+      const builtPath = join(OUTPUT_DIR, 'api-client', fileName);
+      const sourcePath = join('api-client', fileName);
+
+      this.state.steps.uploadFiles.output.builtFileInfos.set(builtPath, {
+        checksum,
+        builtPath,
+        sourcePath,
+        fileFolder: FileFolder.Dependencies,
+      });
+
+      this.uploadFile(builtPath, sourcePath, FileFolder.Dependencies);
+    }
   }
 
   private uploadPendingFiles(): void {
